+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_file_system_client.hpp
@@ -93,49 +93,58 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     /**
      * @brief Creates the file system.
      * @param options Optional parameters to create this file system.
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<Models::CreateDataLakeFileSystemResult> containing the
      * information of create a file system.
      * @remark This request is sent to blob endpoint.
      */
     Azure::Core::Response<Models::CreateDataLakeFileSystemResult> Create(
-        const CreateDataLakeFileSystemOptions& options = CreateDataLakeFileSystemOptions()) const;
+        const CreateDataLakeFileSystemOptions& options = CreateDataLakeFileSystemOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Creates the file system if it does not exists.
      * @param options Optional parameters to create this file system.
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<Models::CreateDataLakeFileSystemResult> containing the
      * information of create a file system. Only valid when successfully created the file system.
      * @remark This request is sent to blob endpoint.
      */
     Azure::Core::Response<Models::CreateDataLakeFileSystemResult> CreateIfNotExists(
-        const CreateDataLakeFileSystemOptions& options = CreateDataLakeFileSystemOptions()) const;
+        const CreateDataLakeFileSystemOptions& options = CreateDataLakeFileSystemOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Deletes the file system.
      * @param options Optional parameters to delete this file system.
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<Models::DeleteDataLakeFileSystemResult> containing the
      * information returned when deleting file systems.
      * @remark This request is sent to blob endpoint.
      */
     Azure::Core::Response<Models::DeleteDataLakeFileSystemResult> Delete(
-        const DeleteDataLakeFileSystemOptions& options = DeleteDataLakeFileSystemOptions()) const;
+        const DeleteDataLakeFileSystemOptions& options = DeleteDataLakeFileSystemOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Deletes the file system if it exists.
      * @param options Optional parameters to delete this file system.
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<Models::DeleteDataLakeFileSystemResult> containing the
      * information returned when deleting file systems. Only valid when successfully deleted the
      * file system.
      * @remark This request is sent to blob endpoint.
      */
     Azure::Core::Response<Models::DeleteDataLakeFileSystemResult> DeleteIfExists(
-        const DeleteDataLakeFileSystemOptions& options = DeleteDataLakeFileSystemOptions()) const;
+        const DeleteDataLakeFileSystemOptions& options = DeleteDataLakeFileSystemOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Sets the metadata of file system.
      * @param metadata User-defined metadata to be stored with the filesystem. Note that the string
      *                 may only contain ASCII characters in the ISO-8859-1 character set.
      * @param options Optional parameters to set the metadata to this file system.
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<Models::SetDataLakeFileSystemMetadataResult> containing the
      * information returned when setting the metadata onto the file system.
      * @remark This request is sent to blob endpoint.
@@ -143,56 +152,64 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     Azure::Core::Response<Models::SetDataLakeFileSystemMetadataResult> SetMetadata(
         Storage::Metadata metadata,
         const SetDataLakeFileSystemMetadataOptions& options
-        = SetDataLakeFileSystemMetadataOptions()) const;
+        = SetDataLakeFileSystemMetadataOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Gets the properties of file system.
      * @param options Optional parameters to get the metadata of this file system.
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<Models::GetDataLakeFileSystemPropertiesResult> containing the
      * information when getting the file system's properties.
      * @remark This request is sent to blob endpoint.
      */
     Azure::Core::Response<Models::GetDataLakeFileSystemPropertiesResult> GetProperties(
         const GetDataLakeFileSystemPropertiesOptions& options
-        = GetDataLakeFileSystemPropertiesOptions()) const;
+        = GetDataLakeFileSystemPropertiesOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief List the paths in this file system.
      * @param recursive If "true", all paths are listed; otherwise, only paths at the root of the
      *                  filesystem are listed.
      * @param options Optional parameters to list the paths in file system.
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<Models::ListPathsSinglePageResult> containing the
      * results when listing the paths under a file system.
      * @remark This request is sent to dfs endpoint.
      */
     Azure::Core::Response<Models::ListPathsSinglePageResult> ListPathsSinglePage(
         bool recursive,
-        const ListPathsSinglePageOptions& options = ListPathsSinglePageOptions()) const;
+        const ListPathsSinglePageOptions& options = ListPathsSinglePageOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Gets the permissions for this file system. The permissions indicate whether
      * file system data may be accessed publicly.
      *
      * @param options Optional parameters to execute this function.
+     * @param context Context for cancelling long running operations.
      * @return A GetDataLakeFileSystemAccessPolicyResult describing the container's access policy.
      * @remark This request is sent to blob endpoint.
      */
     Azure::Core::Response<Models::GetDataLakeFileSystemAccessPolicyResult> GetAccessPolicy(
         const GetDataLakeFileSystemAccessPolicyOptions& options
-        = GetDataLakeFileSystemAccessPolicyOptions()) const;
+        = GetDataLakeFileSystemAccessPolicyOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Sets the permissions for the specified file system. The permissions indicate
      * whether file system's data may be accessed publicly.
      *
-     * @param options Optional
-     * parameters to execute this function.
+     * @param options Optional parameters to execute this function.
+     * @param context Context for cancelling long running operations.
      * @return A SetDataLakeFileSystemAccessPolicyResult describing the updated file system.
      * @remark This request is sent to blob endpoint.
      */
     Azure::Core::Response<Models::SetDataLakeFileSystemAccessPolicyResult> SetAccessPolicy(
         const SetDataLakeFileSystemAccessPolicyOptions& options
-        = SetDataLakeFileSystemAccessPolicyOptions()) const;
+        = SetDataLakeFileSystemAccessPolicyOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Renames a file. By default, the destination is overwritten and
@@ -200,13 +217,15 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
      * @param fileName The file that gets renamed.
      * @param destinationFilePath The path of the file the source file is renaming to.
      * @param options Optional parameters to rename a file
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<DataLakeFileClient> The client targets the renamed file.
      * @remark This request is sent to dfs endpoint.
      */
     Azure::Core::Response<DataLakeFileClient> RenameFile(
         const std::string& fileName,
         const std::string& destinationFilePath,
-        const RenameDataLakeFileOptions& options = RenameDataLakeFileOptions()) const;
+        const RenameDataLakeFileOptions& options = RenameDataLakeFileOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
      * @brief Renames a directory. By default, the destination is overwritten and
@@ -214,6 +233,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
      * @param directoryName The directory that gets renamed.
      * @param destinationDirectoryPath The destinationPath the source directory is renaming to.
      * @param options Optional parameters to rename a directory.
+     * @param context Context for cancelling long running operations.
      * @return Azure::Core::Response<DataLakeDirectoryClient> The client targets the renamed
      * directory.
      * @remark This request is sent to dfs endpoint.
@@ -221,7 +241,8 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     Azure::Core::Response<DataLakeDirectoryClient> RenameDirectory(
         const std::string& directoryName,
         const std::string& destinationDirectoryPath,
-        const RenameDataLakeDirectoryOptions& options = RenameDataLakeDirectoryOptions()) const;
+        const RenameDataLakeDirectoryOptions& options = RenameDataLakeDirectoryOptions(),
+        const Azure::Core::Context& context = Azure::Core::Context()) const;
 
   private:
     Azure::Core::Http::Url m_fileSystemUrl;