+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes
 
 - Removed `BreakBlobLeaseResult::Leasetime`.
+- Moved `Azure::Core::Context` out of options bag of each API, and make it the last optional parameter.
 
 ## 12.0.0-beta.7 (2021-02-03)