+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -31,7 +31,7 @@
 - Renamed `FileGetRangeListResult::FileContentLength` to `FileSize`.
 - Renamed `StorageServiceProperties` to `FileServiceProperties`.
 - Removed `LeaseTime` in results returned by lease operations. Also removed `LeaseId` in `ShareBreakLeaseResult`.
-
+- Moved `Azure::Core::Context` out of options bag of each API, and make it the last optional parameter.
 
 ## 12.0.0-beta.7 (2021-02-04)